Plotext can't correctly handle uint16
This took me a long time. I thought that my handling of the data was incorrect and that I apparently did not understand data types and endianness. There was a large difference in the plots for array.array and numpy.array conversion. Turns out that it was plotext which just could not handle uint16 types correctly...